refactor: :recycle: add namespace Occupations
